We store references to hostname, userid, virtual_host, and password inside a RabbitMQ state object. However, once those arguments leave scope inside the init() function, Python deallocates the memory.
If we want to keep references to them, we need to do custom malloc operations
to store them and deallocate them when we're no longer using the object.

I noticed this problem when putting a breakpoint on the
kombu/transport/librabbitmq.py establish_connection() and attempted to print
the Connection object. On multiple publishing, I noticed the Connection string
object corrupted, particularly the hostname.