body_size should not be a pointer here.
This was breaking body_size encoding in the content frame on big endian servers in 64 bit. Unfortunately, it seems PyArg_ParseTuple t# will only write to a 32 bit integer, not a 64 bit integer, at least on Python 2.7.10 and the docs show the same for future releases.  PyRabbitMQ_Connection_basic_publish is limited compared to the the AMQP spec which allows for 64-bit body sizes.